+++ ems/src/main/java/com/expeditors/ems/repository/HiringStatusRepository.java
@@ -3,6 +3,6 @@ package com.expeditors.ems.repository;
 import com.expeditors.ems.entity.HiringStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HiringStatusRepository extends JpaRepository<HiringStatus,Integer> {
+public interface HiringStatusRepository extends JpaRepository<HiringStatus,Long> {
 
 }
